MTDSA-8106 Retrieve UK Prop BSAS alignments

+++ app/v2/models/response/retrieveBsas/Profit.scala
@@ -20,12 +20,12 @@ import play.api.libs.functional.syntax._
 import play.api.libs.json._
 
 case class Profit(net: Option[BigDecimal],
-                  taxable: Option[BigDecimal])
+                  taxable: Option[BigInt])
 
 object Profit {
   implicit val reads: Reads[Profit] = (
     (JsPath \ "netProfit").readNullable[BigDecimal] and
-      (JsPath \ "taxableProfit").readNullable[BigDecimal]
+      (JsPath \ "taxableProfit").readNullable[BigInt]
     )(Profit.apply _)
 
   implicit val writes: OWrites[Profit] = Json.writes[Profit]
